[HTTP] Use std::find_if instead of std::remove_if
I was confused by the name, but std::find_if finds one element (or not) and returns the iterator (or std::end). We can use this with std::vector::erase to erase it from the vector. std::remove_if does some shifting of values, and std::remove_if is useful for multiple objects to be erased. I only want to erase 1 item.